+++ src/components/AdminSettings.vue
@@ -1,103 +1,29 @@
 <template>
-	<div id="welcome_prefs" class="section">
+	<div id="htmlwidget_prefs" class="section">
 		<h2>
-			<a class="icon icon-welcome" />
-			{{ t('welcome', 'Welcome widget') }}
+			<a class="icon icon-htmlwidget" />
+			{{ t('htmlwidget', 'HtmlWidget') }}
 		</h2>
-		<p class="settings-hint">
-			<span class="icon icon-details" />
-			{{ t('welcome', 'The dashboard welcome widget will be displayed for all users only if you choose a markdown file.') }}
-		</p>
 		<div class="grid-form">
-			<label for="welcome-file-path">
-				<span class="icon icon-file" />
-				{{ t('welcome', 'Markdown content file') }}
+			<label for="htmlwidget-widget-title">
+				<span class="icon icon-tag" />
+				{{ t('htmlwidget', 'Widget Title') }}
 			</label>
-			<div>
-				<button @click="selectFile">
-					<span class="icon icon-folder" />
-				</button>
-				<input id="welcome-file-path"
-					type="text"
-					:value="fullFilePath"
-					:readonly="true"
-					:placeholder="t('welcome', 'No file')"
-					@click="selectFile">
-				<button v-if="state.filePath"
-					@click="clear">
-					<span class="icon icon-delete" />
-				</button>
-			</div>
-		</div>
-		<br>
-		<div v-if="state.filePath"
-			class="grid-form">
-			<label for="welcome-support">
-				<span class="icon icon-user" />
-				{{ t('welcome', 'Support contact') }}
-			</label>
-			<div v-if="state.supportUserId">
-				<Avatar
-					:size="40"
-					:user="state.supportUserId"
-					:tooltip-message="state.supportUserName" />
-				<span class="support-user-name">
-					{{ state.supportUserName }}
-				</span>
-				<button
-					@click="clearSupportContact">
-					<span class="icon icon-delete" />
-				</button>
-			</div>
-			<div v-else>
-				<Multiselect
-					ref="multiselect"
-					class="support-input"
-					label="displayName"
-					:clear-on-select="false"
-					:hide-selected="false"
-					:internal-search="false"
-					:loading="loadingUsers"
-					:options="formattedSuggestions"
-					:placeholder="t('welcome', 'Choose a support user')"
-					:preselect-first="false"
-					:preserve-search="true"
-					:searchable="true"
-					:user-select="true"
-					@search-change="asyncFind"
-					@select="supportContactSelected">
-					<template #option="{option}">
-						<Avatar
-							class="support-avatar-option"
-							:user="option.user"
-							:show-user-status="false" />
-						<span>
-							{{ option.displayName }}
-						</span>
-					</template>
-					<template #noOptions>
-						{{ t('welcome', 'No recommendations. Start typing.') }}
-					</template>
-					<template #noResult>
-						{{ t('welcome', 'No result.') }}
-					</template>
-				</Multiselect>
-			</div>
-			<label for="welcome-support-text">
-				<span class="icon icon-file" />
-				{{ t('welcome', 'Support text') }}
-			</label>
-			<input id="welcome-support-text"
-				v-model="state.supportText"
+			<input id="htmlwidget-widget-title"
+				v-model="state.widgetTitle"
 				type="text"
 				:class="{ 'icon-loading-small': saving }"
-				:placeholder="t('welcome', 'Example: Call {name} to get help.')"
-				@input="onSupportTextChange">
+				@input="onWidgetTitleChange">
+			<label for="htmlwidget-content-html">
+				<span class="icon icon-timezone" />
+				{{ t('htmlwidget', 'HTML') }}
+			</label>
+			<input id="htmlwidget-content-html"
+				v-model="state.contentHtml"
+				type="text"
+				:class="{ 'icon-loading-small': saving }"
+				@input="onContentHtmlChange">
 			<div />
-			<span class="settings-hint">
-				<span class="icon icon-details" />
-				{{ t('welcome', '{name} will be replaced by the support user name') }}
-			</span>
 		</div>
 	</div>
 </template>
@@ -108,8 +34,6 @@ import { generateUrl, generateOcsUrl } from '@nextcloud/router'
 import axios from '@nextcloud/axios'
 import { getCurrentUser } from '@nextcloud/auth'
 import { showSuccess, showError } from '@nextcloud/dialogs'
-import Multiselect from '@nextcloud/vue/dist/Components/Multiselect'
-import Avatar from '@nextcloud/vue/dist/Components/Avatar'
 
 import { delay } from '../utils'
 
@@ -117,15 +41,13 @@ export default {
 	name: 'AdminSettings',
 
 	components: {
-		Multiselect,
-		Avatar,
 	},
 
 	props: [],
 
 	data() {
 		return {
-			state: loadState('welcome', 'admin-config'),
+			state: loadState('htmlwidget', 'admin-config'),
 			saving: false,
 			currentUser: getCurrentUser(),
 			query: '',
@@ -135,34 +57,6 @@ export default {
 	},
 
 	computed: {
-		fullFilePath() {
-			return this.state.filePath
-				? this.state.userName + this.state.filePath
-				: ''
-		},
-		formattedSuggestions() {
-			const result = this.suggestions.map((s) => {
-				return {
-					user: s.id,
-					displayName: s.label,
-					icon: 'icon-user',
-					multiselectKey: s.id + s.label,
-				}
-			})
-			if (this.currentUser) {
-				const lowerCurrent = this.currentUser.displayName.toLowerCase()
-				const lowerQuery = this.query.toLowerCase()
-				if (this.query === '' || lowerCurrent.match(lowerQuery)) {
-					result.push({
-						user: this.currentUser.uid,
-						displayName: this.currentUser.displayName,
-						icon: 'icon-user',
-						multiselectKey: this.currentUser.uid + this.currentUser.displayName,
-					})
-				}
-			}
-			return result
-		},
 	},
 
 	watch: {
@@ -177,14 +71,14 @@ export default {
 			const req = {
 				values,
 			}
-			const url = generateUrl('/apps/welcome/admin-config')
+			const url = generateUrl('/apps/htmlwidget/admin-config')
 			axios.put(url, req)
 				.then((response) => {
-					showSuccess(t('welcome', 'Welcome admin options saved'))
+					showSuccess(t('htmlwidget', 'HtmlWidget admin options saved'))
 				})
 				.catch((error) => {
 					showError(
-						t('welcome', 'Failed to save welcome admin options')
+						t('htmlwidget', 'Failed to save htmlwidget admin options')
 						+ ': ' + (error.response?.request?.responseText ?? '')
 					)
 					console.debug(error)
@@ -192,34 +86,6 @@ export default {
 				.then(() => {
 					this.saving = false
 				})
-		},
-		clear() {
-			this.state.filePath = ''
-			this.state.userName = ''
-			this.state.userId = ''
-			this.saveOptions({
-				filePath: '',
-				userName: '',
-				userId: '',
-			})
-		},
-		selectFile() {
-			OC.dialogs.filepicker(
-				t('welcome', 'Choose markdown welcome content file'),
-				(targetPath) => {
-					this.state.filePath = targetPath
-					this.state.userName = this.currentUser.displayName
-					this.state.userId = this.currentUser.uid
-					this.saveOptions({
-						filePath: this.state.filePath,
-						userName: this.state.userName,
-						userId: this.state.userId,
-					})
-				},
-				false,
-				['text/markdown'],
-				true
-			)
 		},
 		asyncFind(query) {
 			this.query = query
@@ -247,26 +113,14 @@ export default {
 				this.loadingUsers = false
 			})
 		},
-		supportContactSelected(user) {
-			console.debug(user)
-			this.state.supportUserId = user.user
-			this.state.supportUserName = user.displayName
-			this.saveOptions({
-				supportUserId: this.state.supportUserId,
-				supportUserName: this.state.supportUserName,
-			})
-		},
-		clearSupportContact() {
-			this.state.supportUserId = ''
-			this.state.supportUserName = ''
-			this.saveOptions({
-				supportUserId: '',
-				supportUserName: '',
-			})
-		},
-		onSupportTextChange() {
+		onContentHtmlChange() {
 			delay(() => {
-				this.saveOptions({ supportText: this.state.supportText })
+				this.saveOptions({ contentHtml: this.state.contentHtml })
+			}, 2000)()
+		},
+		onWidgetTitleChange() {
+			delay(() => {
+				this.saveOptions({ widgetTitle: this.state.widgetTitle })
 			}, 2000)()
 		},
 	},
@@ -274,12 +128,12 @@ export default {
 </script>
 
 <style scoped lang="scss">
-#welcome_prefs .icon {
+#htmlwidget_prefs .icon {
 	display: inline-block;
 	width: 32px;
 }
 
-#welcome_prefs .grid-form .icon {
+#htmlwidget_prefs .grid-form .icon {
 	margin-bottom: -3px;
 }
 
@@ -323,14 +177,14 @@ export default {
 	margin-right: 10px;
 }
 
-.icon-welcome {
+.icon-htmlwidget {
 	background-image: url(./../../img/app-dark.svg);
 	background-size: 23px 23px;
 	height: 23px;
 	margin-bottom: -4px;
 }
 
-body.theme--dark .icon-welcome {
+body.theme--dark .icon-htmlwidget {
 	background-image: url(./../../img/app.svg);
 }