+++ webpack.js
@@ -6,13 +6,13 @@ const isDev = buildMode === 'development'
 webpackConfig.devtool = isDev ? 'cheap-source-map' : 'source-map'
 
 webpackConfig.stats = {
-    colors: true,
-    modules: false,
+	colors: true,
+	modules: false,
 }
 
 webpackConfig.entry = {
-    adminSettings: { import: path.join(__dirname, 'src', 'adminSettings.js'), filename: 'welcome-adminSettings.js' },
-    dashboard: { import: path.join(__dirname, 'src', 'dashboard.js'), filename: 'welcome-dashboard.js' },
+	adminSettings: {import: path.join(__dirname, 'src', 'adminSettings.js'), filename: 'htmlwidget-adminSettings.js'},
+	dashboard: {import: path.join(__dirname, 'src', 'dashboard.js'), filename: 'htmlwidget-dashboard.js'},
 }
 
 module.exports = webpackConfig